add welcome page from which one can go to 2 sub pages
