+++ package.json
@@ -33,7 +33,8 @@
     "d3": "^7.9.0",
     "eslint": "^7.32.0",
     "eslint-plugin-vue": "^8.7.1",
-    "vue-eslint-parser": "^9.4.3"
+    "vue-eslint-parser": "^9.4.3",
+    "vue-router": "^4.0.13"
   },
   "eslintConfig": {
     "root": true,
