+++ src/App.vue
@@ -1,44 +1,9 @@
 <template>
-  <div id="app">
-    <CSSGrid
-      :selectedProjectId="selectedProjectId"
-      :selectedProjectName="selectedProjectName"
-      @update:selectedProject="handleUpdateSelectedProject"
-    />
-  </div>
+  <router-view />
 </template>
 
 <script>
-import CSSGrid from './components/CSSGrid.vue';
-
 export default {
-  components: {
-    CSSGrid
-  },
-  data() {
-    return {
-      selectedProjectId: null,
-      selectedProjectName: null,
-    };
-  },
-  methods: {
-    handleUpdateSelectedProject(newProjectId, newProjectName) {
-      console.log(`received 'update:selectedProject(${newProjectId}, ${newProjectName})'`);
-      this.selectedProjectId = newProjectId;
-      this.selectedProjectName = newProjectName;
-    }
-  }
+  name: 'RootApp',
 };
 </script>
-
-<style>
-/* Global styles (if any) */
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 40px;
-}
-</style>
