+++ src/components/VirtualDataInfrastructure.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    <h1>Virtual Data Infrastructure</h1>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'VirtualDataInfrastructure',
+};
+</script>
